Moved new Philips motion sensors class into motion.py, added outdoor SML004 sensor
