+++ zhaquirks/philips/motion.py
@@ -26,7 +26,7 @@ from zhaquirks.const import (
     OUTPUT_CLUSTERS,
     PROFILE_ID,
 )
-from zhaquirks.philips import PHILIPS, OccupancyCluster
+from zhaquirks.philips import PHILIPS, SIGNIFY, OccupancyCluster
 
 
 class PhilipsMotion(CustomDevice):
@@ -101,6 +101,57 @@ class PhilipsMotion(CustomDevice):
                     OccupancyCluster,
                 ],
                 OUTPUT_CLUSTERS: [Ota.cluster_id],
+            },
+        }
+    }
+
+
+class SignifyMotion(CustomDevice):
+    """New Philips motion sensor devices."""
+
+    signature = {
+        MODELS_INFO: [(SIGNIFY, "SML003"), (SIGNIFY, "SML004")],
+        ENDPOINTS: {
+            2: {
+                PROFILE_ID: zha.PROFILE_ID,
+                DEVICE_TYPE: zha.DeviceType.OCCUPANCY_SENSOR,
+                INPUT_CLUSTERS: [
+                    Basic.cluster_id,
+                    PowerConfiguration.cluster_id,
+                    Identify.cluster_id,
+                    IlluminanceMeasurement.cluster_id,
+                    TemperatureMeasurement.cluster_id,
+                    OccupancySensing.cluster_id,
+                ],
+                OUTPUT_CLUSTERS: [
+                    Basic.cluster_id,
+                    Identify.cluster_id,
+                    OnOff.cluster_id,
+                    Ota.cluster_id,
+                ],
+            },
+        },
+    }
+
+    replacement = {
+        ENDPOINTS: {
+            2: {
+                PROFILE_ID: zha.PROFILE_ID,
+                DEVICE_TYPE: zha.DeviceType.OCCUPANCY_SENSOR,
+                INPUT_CLUSTERS: [
+                    Basic.cluster_id,
+                    PowerConfiguration.cluster_id,
+                    Identify.cluster_id,
+                    IlluminanceMeasurement.cluster_id,
+                    TemperatureMeasurement.cluster_id,
+                    OccupancyCluster,
+                ],
+                OUTPUT_CLUSTERS: [
+                    Basic.cluster_id,
+                    Identify.cluster_id,
+                    OnOff.cluster_id,
+                    Ota.cluster_id,
+                ],
             },
         }
     }
